upd + delete files + flux/feed
