+++ admin/inc/addons.php
@@ -13,7 +13,7 @@
 
 
 // temp, put this include somewhere else
-include 'file-system.php';
+include 'filesystem.php';
 
 /**
  *
